Restructured the project and remade cmake files
