+++ src/algorithm/algorithm_functions.c
