+++ src/algorithm/algorithm_functions.h
