+++ src/dataparser/CMakeLists.txt
@@ -1,2 +1,7 @@
-add_library(csv_dataparser csv_parser.c)
-target_include_directories(csv_dataparser PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+add_library(dataparser
+        csv_parser.c
+)
+
+target_include_directories(dataparser PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+
+target_link_libraries(dataparser PUBLIC graph)
