+++ src/graph/graph_functions.c
