+++ src/graph/graph_functions.h
