+++ src/sample_network/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(sample_network
+        sample_network.c
+)
+
+target_include_directories(sample_network PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
